# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  contract.OrderAddonCreateReq:
    properties:
      category:
        description: 分类 makeup-妆造 urgency-时效 service-服务 retouch-精修
        type: string
      confirmed:
        description: 客户是否确认 0-待确认 1-已确认
        type: integer
      name:
        description: 加项名称
        type: string
      order_id:
        description: 订单ID（body 主键）
        type: integer
      price:
        description: 单价
        type: number
      qty:
        description: 数量（<=0 按 1 计）
        type: integer
      remark:
        description: 备注
        type: string
    required:
    - name
    type: object
...
  contract.OrderCancelReq:
    properties:
      reason:
        description: 取消原因
        type: string
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      trace_id:
        description: 当前请求链路 trace_id，便于前端按链路反馈问题
        type: string
    type: object
  response.Page:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: |-
    摄影工作室管理系统后端 API。五端挂载：pc（无前缀）/ miniapp（/miniapp）/ h5（/h5）/ wechat 客户区（/wechat）/ 员工端（/wechat/staff）。
    约定：业务接口统一使用 POST + JSON body（含分页 page/page_size、keyword 等），路径参数用 :id / :order_id；统一响应体 response.Body{code,msg,data,trace_id}，列表数据为 response.Page{list,total,page,page_size}。
    鉴权：除登录、健康检查与客户区公开接口（ClientPublic）外，均需在 Authorization 头携带 Bearer <token>（员工 JWT 或客户 JWT）。
  title: Photography Server API
  version: "1.0"
paths:
...
  /order/addon/create:
    post:
      consumes:
      - application/json
      description: 新增加项并在同一事务内重算订单金额；body.order_id 为订单主键。
      parameters:
      - description: 加项信息（含 order_id）
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.OrderAddonCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderAddon'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 新建订单加项
      tags:
      - 订单加项
  /order/addon/delete/{id}:
    post:
      consumes:
      - application/json
      description: 删除加项并在同一事务内重算订单金额。
      parameters:
      - description: 加项ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除订单加项
      tags:
      - 订单加项
  /order/addon/list/{order_id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 订单ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.OrderAddon'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 订单加项列表
      tags:
      - 订单加项
  /order/addon/update:
    post:
      consumes:
      - application/json
      description: 更新加项并在同一事务内重算订单金额；body.id 为加项主键。
      parameters:
      - description: 加项信息（含 id）
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.OrderAddonUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderAddon'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新订单加项
      tags:
      - 订单加项
  /order/cancel/{id}:
    post:
      consumes:
      - application/json
      description: 取消订单并记录原因；当前状态不允许取消时返回业务错误。
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 取消原因
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.OrderCancelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 取消订单
      tags:
      - 订单
  /order/create:
    post:
      consumes:
      - application/json
      description: 创建订单；客户与线索二选一（customer_id / lead_id），套餐必填。
      parameters:
      - description: 订单信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.OrderCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建订单
      tags:
      - 订单
  /order/detail/{id}:
    post:
      consumes:
      - application/json
      description: 返回订单及其收款、退款、日志、交付与可流转状态。
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/contract.OrderDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 订单详情
      tags:
      - 订单
  /order/list:
    post:
      consumes:
      - application/json
      description: 分页查询订单，可按状态、客户过滤。
      parameters:
      - description: 查询条件
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.OrderListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.Page'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.Order'
                        type: array
                      page:
                        type: integer
                      page_size:
                        type: integer
                      total:
                        type: integer
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 订单列表
      tags:
      - 订单
  /order/logs/{id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.OrderLog'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 订单操作日志
      tags:
      - 订单
  /order/reschedule/apply/{order_id}:
    post:
      consumes:
      - application/json
      description: 管理端代客户发起改期申请（改期统一走改期单链路，不直接改订单拍摄日期）。
      parameters:
      - description: 订单ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: 改期请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.RescheduleApplyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderReschedule'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 发起改期
      tags:
      - 改期
  /order/reschedule/audit/{id}:
    post:
      consumes:
      - application/json
      description: 审批改期申请；approved 为 true 通过、false 驳回，缺省返回参数错误。
      parameters:
      - description: 改期单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 审批请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.AuditReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 改期审批
      tags:
      - 改期
  /order/reschedule/list/{order_id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 订单ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.OrderReschedule'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 订单改期单列表
      tags:
      - 改期
...
  /order/update:
    post:
      consumes:
      - application/json
      description: 更新订单拍摄信息；body.id 为主键。
      parameters:
      - description: 订单信息（含 id）
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/contract.OrderUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新订单
      tags:
      - 订单
...
securityDefinitions:
  BearerAuth:
    description: 先调用登录接口获取 token，再以 "Bearer <token>" 形式放入 Authorization 头
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"